Refactoring code, move from Main to separate files
