+++ .vscode/settings.json
@@ -18,5 +18,6 @@
         "string_view": "cpp",
         "vector": "cpp"
     },
-    "mesonbuild.configureOnOpen": false
+    "mesonbuild.configureOnOpen": false,
+    "C_Cpp.errorSquiggles": "disabled"
 }
