+++ include/GameSystem.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "PlayerControls.h"
+#include "Background.h"
+#include "Platform.h"
