+++ meson.build
@@ -31,7 +31,10 @@ sources = [
     'src/Rendering.cpp',
     'src/SpriteRendering.cpp',
     'src/Windowing.cpp',
-    'src/Main.cpp'
+    'src/Main.cpp',
+    'src/PlayerControls.cpp',
+    'src/Platform.cpp',
+    'src/Background.cpp'
 ]
 
 executable('dagger',
